refactor: review helper
