+++ test/destroy.test.js
@@ -2,7 +2,7 @@
 
 const fs = require('fs')
 const SonicBoom = require('../')
-const { file, runTests } = require('./helper')
+const { file, runTests, once } = require('./helper')
 
 runTests(buildTests)
 
@@ -10,40 +10,38 @@ function buildTests (test, sync) {
   // Reset the umask for testing
   process.umask(0o000)
 
-  test('destroy', (t) => {
-    t.plan(5)
+  test('destroy', async (t) => {
+    t.plan(4)
 
     const dest = file()
     const fd = fs.openSync(dest, 'w')
     const stream = new SonicBoom({ fd, sync })
 
-    t.ok(stream.write('hello world\n'))
+    t.assert.ok(stream.write('hello world\n'))
     stream.destroy()
-    t.throws(() => { stream.write('hello world\n') })
+    t.assert.throws(() => { stream.write('hello world\n') })
 
-    fs.readFile(dest, 'utf8', function (err, data) {
-      t.error(err)
-      t.equal(data, 'hello world\n')
-    })
+    const data = fs.readFileSync(dest, 'utf8')
+    t.assert.strictEqual(data, 'hello world\n')
 
     stream.on('finish', () => {
-      t.fail('finish emitted')
+      t.assert.fail('finish emitted')
     })
 
-    stream.on('close', () => {
-      t.pass('close emitted')
+    await once(stream, 'close', () => {
+      t.assert.ok('close emitted')
     })
   })
 
-  test('destroy while opening', (t) => {
+  test('destroy while opening', async (t) => {
     t.plan(1)
 
     const dest = file()
     const stream = new SonicBoom({ dest })
 
     stream.destroy()
-    stream.on('close', () => {
-      t.pass('close emitted')
+    await once(stream, 'close', () => {
+      t.assert.ok('close emitted')
     })
   })
 }
