+++ test/helper.js
@@ -1,6 +1,7 @@
 'use strict'
 
-const { test, teardown } = require('tap')
+const tap = require('tap')
+const nodeTest = require('node:test')
 const fs = require('fs')
 const os = require('os')
 const path = require('path')
@@ -14,7 +15,7 @@ function file () {
   return file
 }
 
-teardown(() => {
+nodeTest.after(() => {
   const rmSync = fs.rmSync || fs.rmdirSync
   files.forEach((file) => {
     try {
@@ -27,16 +28,42 @@ teardown(() => {
   })
 })
 
-function runTests (buildTests) {
-  test('sync false', (t) => {
+function runTestsLegacy (buildTests) {
+  tap.test('sync false', (t) => {
     buildTests(t.test, false)
     t.end()
   })
 
-  test('sync true', (t) => {
+  tap.test('sync true', (t) => {
     buildTests(t.test, true)
     t.end()
   })
 }
 
-module.exports = { file, runTests }
+async function runTests (buildTests) {
+  nodeTest.describe('sync false', () => {
+    buildTests(nodeTest.test, false)
+  })
+
+  nodeTest.describe('sync true', () => {
+    buildTests(nodeTest.test, true)
+  })
+}
+
+/**
+ * Listens for an event on an object and resolves a promise when the event is emitted.
+ * @param {Object} emitter - The object to listen to.
+ * @param {string} event - The name of the event to listen for.
+ * @param {Function} fn - The function to call when the event is emitted.
+ * @returns {Promise} A promise that resolves when the event is emitted.
+ */
+function once (emitter, event, fn) {
+  return new Promise(resolve => {
+    emitter.on(event, (...args) => {
+      fn(...args)
+      resolve()
+    })
+  })
+}
+
+module.exports = { file, runTestsLegacy, runTests, once }
