+++ test/write.test.js
@@ -4,7 +4,7 @@ const { test } = require('tap')
 const fs = require('fs')
 const proxyquire = require('proxyquire')
 const SonicBoom = require('../')
-const { file, runTests } = require('./helper')
+const { file, runTestsLegacy: runTests } = require('./helper')
 
 runTests(buildTests)
